# lib/api-zod/src/generated/types/appSettings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Study Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface AppSettings {
  id: number;
  theme: string;
  updatedAt: string;
}