# lib/api-zod/src/generated/types/noteUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Study Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface NoteUpdate {
  classLabel?: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  subjectId?: number | null;
}