# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Study Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface Resource {
  id: number;
  subjectName: string;
  topicName: string;
  url: string;
  isPaid: boolean;
  createdAt: string;
}