# lib/api-zod/src/generated/types/routineItemUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Study Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface RoutineItemUpdate {
  day?: string;
  /** @nullable */
  subjectId?: number | null;
  label?: string;
}