# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Study Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskInput {
  /** @nullable */
  subjectId?: number | null;
  name: string;
  date: string;
  completed?: boolean;
}