# lib/api-zod/src/generated/types/topic.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Study Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface Topic {
  id: number;
  subjectId: number;
  name: string;
  completed: boolean;
  createdAt: string;
}