# lib/api-zod/src/generated/types/topicInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Study Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface TopicInput {
  subjectId: number;
  name: string;
  completed?: boolean;
}